Haskell - Day06 done
Using Array data type made things easier (no need to propagate the
limit parameter)

Still went the naive way (for each new visited position, we turn it
into an obstacle then check if it is looping.

It is quite slow (6 secs + when compiled vs. 88secs in GHCI :| )

One way to speed it up would be to get an ordered  Set of all
obstacles, and keeping track of the turns (each obstacle would require
a turn before it) if we detect a turn we already did it means we are
in a loop.